refactor: AuthService interface / implementation으로 변경
